Move useSessionParams
