+++ frontend/src/__generated/index.css
@@ -5731,6 +5731,67 @@ body {
   height: 20px;
   pointer-events: none;
 }
+._1ek953u0 {
+  z-index: 99999;
+  height: 100vh;
+  width: 100vw;
+  display: flex;
+  justify-content: center;
+  position: fixed;
+  top: 0;
+  left: 0;
+}
+._1ek953u1 {
+  background-color: rgba(111, 110, 119, 0.48);
+  z-index: 99999;
+}
+._1ek953u2 {
+  width: 580px;
+  top: 25vh;
+}
+._1ek953u3 {
+  width: 100%;
+}
+._1ek953u4 {
+  flex-shrink: 0;
+}
+._1ek953u5 {
+  flex-shrink: 0;
+}
+._1ek953u6 {
+  border-top-right-radius: 0;
+  border-bottom-right-radius: 0;
+}
+._1ek953u7 {
+  border-top-left-radius: 0;
+  border-bottom-left-radius: 0;
+}
+._1ek953u8 {
+  flex: 1;
+}
+._1ek953u9:hover {
+  cursor: pointer;
+}
+.bsyaa20 {
+  text-decoration: none;
+  color: var(--_1pyqka9y);
+}
+.bsyaa20:hover {
+  color: var(--_1pyqka9y);
+}
+._1kpm5g60 {
+  text-decoration: none;
+}
+._1kpm5g60:focus-visible {
+  outline: revert !important;
+}
+._1vb6x0p0 {
+  text-decoration: none;
+  color: var(--_1pyqka9y);
+}
+._1vb6x0p0:hover {
+  color: var(--_1pyqka9y);
+}
 .ltrwrc0 {
   position: fixed;
   left: 100%;
@@ -5848,67 +5909,6 @@ body {
 }
 ._1gpgayif {
   filter: blur(4px);
-}
-._1ek953u0 {
-  z-index: 99999;
-  height: 100vh;
-  width: 100vw;
-  display: flex;
-  justify-content: center;
-  position: fixed;
-  top: 0;
-  left: 0;
-}
-._1ek953u1 {
-  background-color: rgba(111, 110, 119, 0.48);
-  z-index: 99999;
-}
-._1ek953u2 {
-  width: 580px;
-  top: 25vh;
-}
-._1ek953u3 {
-  width: 100%;
-}
-._1ek953u4 {
-  flex-shrink: 0;
-}
-._1ek953u5 {
-  flex-shrink: 0;
-}
-._1ek953u6 {
-  border-top-right-radius: 0;
-  border-bottom-right-radius: 0;
-}
-._1ek953u7 {
-  border-top-left-radius: 0;
-  border-bottom-left-radius: 0;
-}
-._1ek953u8 {
-  flex: 1;
-}
-._1ek953u9:hover {
-  cursor: pointer;
-}
-.bsyaa20 {
-  text-decoration: none;
-  color: var(--_1pyqka9y);
-}
-.bsyaa20:hover {
-  color: var(--_1pyqka9y);
-}
-._1kpm5g60 {
-  text-decoration: none;
-}
-._1kpm5g60:focus-visible {
-  outline: revert !important;
-}
-._1vb6x0p0 {
-  text-decoration: none;
-  color: var(--_1pyqka9y);
-}
-._1vb6x0p0:hover {
-  color: var(--_1pyqka9y);
 }
 ._14ud0dc0::-webkit-scrollbar {
   width: 9px;
