+++ frontend/src/pages/Player/RightPlayerPanel/components/SessionInsights/SessionInsights.tsx
@@ -13,7 +13,7 @@ import { Button } from '@/components/Button'
 import { useGetSessionInsightLazyQuery } from '@/graph/generated/hooks'
 import usePlayerConfiguration from '@/pages/Player/PlayerHook/utils/usePlayerConfiguration'
 import { useReplayerContext } from '@/pages/Player/ReplayerContext'
-import { useSessionParams } from '@/pages/Player/utils/utils'
+import { useSessionParams } from '@/pages/Sessions/utils'
 import { useGlobalContext } from '@/routers/ProjectRouter/context/GlobalContext'
 import { playerTimeToSessionAbsoluteTime } from '@/util/session/utils'
 import { MillisToMinutesAndSeconds } from '@/util/time'
